# docs/swagger.yaml
basePath: /api
definitions:
  api.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      code:
        description: Error code for client-side error handling
        type: string
      details:
        description: Additional details for debugging
        type: string
      error:
        description: User-friendly error message
        type: string
      request_id:
        description: Request ID for tracing
        type: string
    type: object
  dto.EventAddressResponse:
    properties:
      address_line:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/dto.UserResponse'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      id:
        type: integer
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/dto.UserResponse'
      updated_by_id:
        type: integer
    type: object
  dto.MealEventAddressResponse:
    properties:
      address:
        $ref: '#/definitions/dto.EventAddressResponse'
      address_id:
        type: integer
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/dto.UserResponse'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      id:
        type: integer
      meal_event_id:
        type: integer
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/dto.UserResponse'
      updated_by_id:
        type: integer
    type: object
  dto.MealEventCreateRequest:
    properties:
      address_ids:
        items:
          type: integer
        type: array
      cutoff_time:
        type: string
      description:
        type: string
      event_date:
        type: string
      event_duration:
        description: in minutes
        minimum: 1
        type: integer
      menu_set_ids:
        items:
          type: integer
        type: array
      name:
        type: string
    required:
    - cutoff_time
    - event_date
    - event_duration
    - name
    type: object
  dto.MealEventResponse:
    properties:
      addresses:
        items:
          $ref: '#/definitions/dto.MealEventAddressResponse'
        type: array
      confirmed_at:
        type: string
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/dto.UserResponse'
      created_by_id:
        type: integer
      cutoff_time:
        type: string
      deleted_at:
        type: string
      description:
        type: string
      event_date:
        type: string
      event_duration:
        description: in minutes
        type: integer
      id:
        type: integer
      menu_sets:
        items:
          $ref: '#/definitions/dto.MealEventSetResponse'
        type: array
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/dto.UserResponse'
      updated_by_id:
        type: integer
    type: object
  dto.MealEventSetResponse:
    properties:
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/dto.UserResponse'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      id:
        type: integer
      label:
        type: string
      meal_event_id:
        type: integer
      menu_set:
        $ref: '#/definitions/dto.MenuSetResponse'
      menu_set_id:
        type: integer
      note:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/dto.UserResponse'
      updated_by_id:
        type: integer
    type: object
  dto.MealEventUpdateRequest:
    properties:
      address_ids:
        items:
          type: integer
        type: array
      cutoff_time:
        type: string
      description:
        type: string
      event_date:
        type: string
      event_duration:
        description: in minutes
        minimum: 1
        type: integer
      menu_set_ids:
        items:
          type: integer
        type: array
      name:
        type: string
    type: object
  dto.MenuItemResponse:
    properties:
      average_rating:
        type: number
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/dto.UserResponse'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      image_url:
        type: string
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/dto.UserResponse'
      updated_by_id:
        type: integer
    type: object
  dto.MenuSetResponse:
    properties:
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/dto.UserResponse'
      created_by_id:
        type: integer
      deleted_at:
        type: string
      description:
        type: string
      id:
        type: integer
      menu_items:
        items:
          $ref: '#/definitions/dto.MenuItemResponse'
        type: array
      name:
        type: string
      updated_at:
        type: string
      updated_by:
        $ref: '#/definitions/dto.UserResponse'
      updated_by_id:
        type: integer
    type: object
...
  dto.UserResponse:
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      department:
        type: string
      email:
        type: string
      employee_id:
        type: string
      id:
        type: integer
      is_active:
        type: boolean
      name:
        type: string
      notification_enabled:
        type: boolean
      role:
        $ref: '#/definitions/model.UserRole'
      updated_at:
        type: string
      username:
        type: string
    type: object
...
  model.UserRole:
    enum:
    - admin
    - employee
    - manager
    type: string
    x-enum-varnames:
    - UserRoleAdmin
    - UserRoleEmployee
    - UserRoleManager
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A meal management system API for employees
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: MealSync API
  version: "1.0"
paths:
...
  /meals:
    get:
      consumes:
      - application/json
      description: Get meal events within a date range
      parameters:
      - description: Start Date (YYYY-MM-DD)
        in: query
        name: start_date
        required: true
        type: string
      - description: End Date (YYYY-MM-DD)
        in: query
        name: end_date
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.MealEventResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List meal events by date range
      tags:
      - meals
    post:
      consumes:
      - application/json
      description: Create a new meal event
      parameters:
      - description: Meal Event Data
        in: body
        name: meal
        required: true
        schema:
          $ref: '#/definitions/dto.MealEventCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.MealEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create meal event
      tags:
      - meals
  /meals/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an existing meal event
      parameters:
      - description: Meal Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/api.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete meal event
      tags:
      - meals
    get:
      consumes:
      - application/json
      description: Get a specific meal event by its ID
      parameters:
      - description: Meal Event ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MealEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get meal event by ID
      tags:
      - meals
    put:
      consumes:
      - application/json
      description: Update an existing meal event
      parameters:
      - description: Meal Event ID
        in: path
        name: id
        required: true
        type: integer
      - description: Meal Event Data
        in: body
        name: meal
        required: true
        schema:
          $ref: '#/definitions/dto.MealEventUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MealEventResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update meal event
      tags:
      - meals
...
security:
- BearerAuth: []
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"